+++ src/tools/blast/types/blastServerParameters.ts
@@ -268,21 +268,6 @@ export type Database =
   | 'uniref50'
   | 'uniref90';
 
-export const databaseToNamespace = (
-  database: Database
-): Namespace.uniprotkb | Namespace.uniparc | Namespace.uniref | undefined => {
-  if (database.startsWith(Namespace.uniprotkb)) {
-    return Namespace.uniprotkb;
-  }
-  if (database.startsWith(Namespace.uniparc)) {
-    return Namespace.uniparc;
-  }
-  if (database.startsWith(Namespace.uniref)) {
-    return Namespace.uniref;
-  }
-  return undefined;
-};
-
 // https://www.ebi.ac.uk/Tools/services/rest/ncbiblast/parameterdetails/hsps
 // HSPs: max number of alignments for each hit; undefined = all
 export type HSPs = undefined | number;
